requirements.txt
altair
pandas
streamlit
streamlit-authenticator
pyyaml